Changes how colormaps are loaded. Adds load and new buttons as well as a save button
